Fixed a large number of WIN64 type conversion warnings.
Changed FS_Write to do multiple write calls if the size exceeds 1GB, this is to accommodate WIN64 where write() takes a UINT rather than size_t.  We have no use of such large writes at this time, so this is a mostly theoretical feature.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12081 d7cf8633-e32d-0410-b094-e92efae38249

+++ cap_avi.c
@@ -473,7 +473,7 @@ static void SCR_CaptureVideo_Avi_SoundFrame(const portable_sampleframe_t *paintb
 		out_ptr[3] = (unsigned char)(n1 >> 8);
 	}
 
-	x = length*4;
+	x = (int)length*4;
 	if(format->canseek)
 	{
 		SCR_CaptureVideo_RIFF_OverflowCheck(8 + x);
