+++ jpeg.c
@@ -970,8 +970,8 @@ size_t JPEG_SaveImage_to_Buffer (char *jpegbuf, size_t jpegsize, int width, int 
 	}
 #endif
 
-	//quality_guess = (100 * jpegsize - 41000) / (width*height) + 2; // fits random data
-	quality_guess   = (256 * jpegsize - 81920) / (width*height) - 8; // fits Nexuiz's/Xonotic's map pictures
+	//quality_guess = (int)((100 * jpegsize - 41000) / (width*height) + 2); // fits random data
+	quality_guess   = (int)((256 * jpegsize - 81920) / (width*height) - 8); // fits Nexuiz's/Xonotic's map pictures
 
 	quality_guess = bound(0, quality_guess, 100);
 	quality = bound(0, quality_guess + sv_writepicture_quality.integer, 100); // assume it can do 10 failed attempts
